# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Agila API
  version: "1.0.0"
  description: >
    Agila API v1. Send Authorization Bearer with a session JWT or an ek_ token.
    The same handlers and permission checks as the web app apply.
    Relay routes add external keys, column titles, claim, and plans.
    OAuth callbacks are not part of this API.
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
...
  "/api/v1/boards/{boardId}/columns":
    get:
      summary: "GET /api/v1/boards/{boardId}/columns"
      description: "Implemented in server/routes/boards.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
...
  "/api/v1/columns":
    post:
      summary: "POST /api/v1/columns"
      description: "Implemented in server/routes/columns.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
  "/api/v1/columns/renumber":
    post:
      summary: "POST /api/v1/columns/renumber"
      description: "Implemented in server/routes/columns.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
  "/api/v1/columns/reorder":
    post:
      summary: "POST /api/v1/columns/reorder"
      description: "Implemented in server/routes/columns.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
  "/api/v1/columns/{id}":
    delete:
      summary: "DELETE /api/v1/columns/{id}"
      description: "Implemented in server/routes/columns.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
    put:
      summary: "PUT /api/v1/columns/{id}"
      description: "Implemented in server/routes/columns.js. Same rules as the web app."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
        '401':
          description: Missing or invalid token
        '403':
          description: Denied by the same rule as the web app, including admin and owner self-delete
        '409':
          description: Conflict, including a stale Relay claim
...